Build rubies with suffix, keg-only when relevant
The rubies are keg-only and installed without a suffix. So one ends up
with:
    /usr/local/Cellar/ruby22/2.2.5/bin/ruby
And a warning about the formula being keg-only preventing symbolic link
for ruby to be put in /usr/local/bin/ruby.

When one wants several rubies, each will typically be build with
--with-suffix and end up with:
    /usr/local/Cellar/ruby22/2.2.5/bin/ruby22

The binaries still have to be manually linked manually. Eg:

    brew install ruby21 --with-suffix
    brew link --force ruby21
    brew install ruby22 --with-suffix
    brew link --force ruby22

Default all rubies to use a suffix unless --without-suffix is passed and
hence no more mark them keg-only.

A gotcha is some of the rubies versions we ship are indeed provided by
Mac OS X. For example 2.0.0 on Sierra, hence when building without a
suffix, mark the recipe as keg-only when the same ruby version is
provided by Apple.

For reference the Mac OS X ruby versions can be obtained on
https://opensource.apple.com/ browsing to the ruby package and looking
at the ruby.plist file.

| Ver | Name | Ruby version
|--|--|--
| 10.12 | sierra | 2.0.0-p648
| 10.11 | el_capitan | 2.0.0-p648
| 10.10 | yosemite | 2.0.0-p481
| 10.9 | mavericks | 2.0.0-p481
| 10.8 | mountain_lion | 1.8.7-p357
| 10.7 | lion | 1.8.7-p357
| 10.6 | snow_leopard | 1.8.7-p174
| 10.5 | leopard | 1.8.6-p287

closes #1357